Rename split_execution.py to read_train_save.py
The file reads data, pre processes the data, defines a deep learning model to predict ratings, saves the model, splits the entire data set in to multiple parts and saves on disk. These parts and the model will be read by prediction file, processed and predicted ratings will be written to multiple files.